Address unicode problems
PR #179 caused some mixing of unicode and strings. Commands like
"ripe-atlas measurement-info 5001" start breaking. This commit tries to
make they way we treat prints and output to file safer. That means we
try to use unicode everywhere and as soon as we print we transform to
byte strings with utf8 encoding. This way outputing to file is safe (it
used to break before).