Fix --location probe search
Google has changed the Geocoding API conditions, now every request has
to be authenticated with an API key of an account connected to a billing
account. Thus I added an option to provide such API key on command line
as well as in the config file.

I also fixed the error reporting so any Google API errors are presented
to the user.

Signed-off-by: Ondřej Caletka <ondrej@caletka.cz>